separate edit-waiting queue dialog
